refactored sample-wrapper to base-wrapper
