+++ base-wrapper/src/main/java/de/dlr/proseo/basewrap/PushedProcessingOutput.java
@@ -1,4 +1,4 @@
-package de.dlr.proseo.samplewrap;
+package de.dlr.proseo.basewrap;
 
 public class PushedProcessingOutput {
 	private long id;
