+++ base-wrapper/src/main/java/de/dlr/proseo/basewrap/package-info.java
@@ -8,4 +8,4 @@
  * @author Dr. Thomas Bassler
  *
  */
-package de.dlr.proseo.samplewrap;
+package de.dlr.proseo.basewrap;
