+++ base-wrapper/src/main/java/de/dlr/proseo/basewrap/rest/RestAuth.java
@@ -1,4 +1,4 @@
-package de.dlr.proseo.samplewrap.rest;
+package de.dlr.proseo.basewrap.rest;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
