+++ base-wrapper/src/main/java/de/dlr/proseo/basewrap/s3/AmazonS3URI.java
@@ -12,7 +12,7 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package de.dlr.proseo.samplewrap.s3;
+package de.dlr.proseo.basewrap.s3;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URI;
