+++ base-wrapper/src/test/resources/s3TestData/file02.txt
