+++ base-wrapper/src/test/resources/s3TestData/file06.txt
